ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
python-multipart==0.0.9
pandas==2.2.0
openpyxl==3.1.2
pdfplumber==0.11.0
python-dotenv==1.0.0
ollama==0.2.0
